Changed Session and Tokens tests
