+++ testsuite/ui/src/test/java/org/keycloak/testsuite/ui/page/settings/TokensPage.java
@@ -12,13 +12,14 @@ import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.ui.Select;
 
 import static java.lang.String.valueOf;
+import static org.keycloak.testsuite.ui.util.SeleniumUtils.waitGuiForElement;
 import static org.apache.commons.lang3.text.WordUtils.capitalize;
 
 /**
  *
  * @author pmensik
  */
-public class SessionAndTokensPage extends AbstractPage {
+public class TokensPage extends AbstractPage {
 	
 	@FindBy(id = "ssoSessionIdleTimeout")
 	private WebElement sessionTimeout;
@@ -32,10 +33,6 @@ public class SessionAndTokensPage extends AbstractPage {
 	@FindBy(name = "ssoSessionMaxLifespanUnit")
 	private Select sessionLifespanTimeoutUnit;
 
-	public void logoutAllSessions() {
-		primaryButton.click();
-	}
-	
 	public void setSessionTimeout(int timeout, TimeUnit unit) {
 		setTimeout(sessionTimeoutUnit, sessionTimeout, timeout, unit);
 	}
@@ -46,6 +43,7 @@ public class SessionAndTokensPage extends AbstractPage {
 	
 	private void setTimeout(Select timeoutElement, WebElement unitElement,
 			int timeout, TimeUnit unit) {
+		waitGuiForElement(sessionTimeout);
 		timeoutElement.selectByValue(capitalize(unit.name().toLowerCase()));
 		unitElement.clear();
 		unitElement.sendKeys(valueOf(timeout));
